+++ NEWS.md
@@ -1,3 +1,16 @@
+# selfcomplexity 0.0.2 _2021-08-20_
+
+  * Update args for `calculate_H` such that both can be used
+  either quoted or unquoted
+  * Add initial test for `calculate_H` function on the simple example
+    * For this I still need verification of what the published reference value is
+  * Flesh out the documentation for `calculate_H` in line with Kasia's info
+  * Add some utility functions for 
+    * Trimming NA values
+    * Extracting unique values from comma separated strings in a vector
+  * Add some sanity checking functionality for inputs to functions
+  * Update DESCRIPTION dependencies to ensure all imports defined.
+
 # selfcomplexity 0.0.2 _2021-08-16_
 
   * Add inteRgrate checks as github action for running lintr and other simple checks